refactor: 의존성 역전 원칙(DIP) 적용
