+++ toby-spring/src/main/java/com/example/tobyspring/ObjectFactory.java
@@ -1,7 +1,7 @@
 package com.example.tobyspring;
 
 import com.example.tobyspring.exchangerate.CachedExchangeRateProvider;
-import com.example.tobyspring.exchangerate.ExchangeRateProvider;
+import com.example.tobyspring.payment.ExchangeRateProvider;
 import com.example.tobyspring.exchangerate.SimpleExchangeRateProvider;
 import com.example.tobyspring.exchangerate.WebApiExchangeRateProvider;
 import com.example.tobyspring.payment.PaymentService;
