+++ toby-spring/src/main/java/com/example/tobyspring/exchangerate/WebApiExchangeRateProvider.java
@@ -1,5 +1,6 @@
 package com.example.tobyspring.exchangerate;
 
+import com.example.tobyspring.payment.ExchangeRateProvider;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.BufferedReader;
 import java.io.IOException;
