+++ toby-spring/src/main/java/com/example/tobyspring/payment/ExchangeRateProvider.java
@@ -1,4 +1,4 @@
-package com.example.tobyspring.exchangerate;
+package com.example.tobyspring.payment;
 
 import java.io.IOException;
 import java.math.BigDecimal;
