+++ toby-spring/src/main/java/com/example/tobyspring/payment/PaymentService.java
@@ -1,6 +1,5 @@
 package com.example.tobyspring.payment;
 
-import com.example.tobyspring.exchangerate.ExchangeRateProvider;
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
